complete...ish?
I was able to complete a version of the certification component that I wanted to make, which included adding bootstrap styles without interfering with styles in the other components. That's done, and it looks good. 
There is another version of the certification component i would like to explore, but that has been tabled until later while i work on a new component...